+++ risk/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.10.0 (Unreleased)
+
+* **Enhancement** Add support for a Bot detection specific data model. [#252](https://github.com/patrickcping/pingone-go-sdk-v2/pull/252)
+
 # v0.9.0 (2023-08-08)
 
 * **Enhancement** Implement HATEOAS links for API response objects. [#227](https://github.com/patrickcping/pingone-go-sdk-v2/pull/227)